+++ src/main/java/org/thermoweb/intellij/plugin/encrypt/Algorithm.java
@@ -1,0 +1,14 @@
+package org.thermoweb.intellij.plugin.encrypt;
+
+public enum Algorithm {
+	PBE("PBEWithHmacSHA512AndAES_128");
+	private final String code;
+
+	Algorithm(final String code) {
+		this.code = code;
+	}
+
+	public String getCode() {
+		return this.code;
+	}
+}