+++ src/main/java/org/thermoweb/intellij/plugin/encrypt/CipherUtils.java
@@ -6,20 +6,23 @@ import org.jasypt.encryption.pbe.config.SimpleStringPBEConfig;
 import org.jasypt.iv.RandomIvGenerator;
 
 public class CipherUtils {
+	private CipherUtils() {
 
-	public String encrypt(final String value, final String password) {
-		return stringEncryptor(password).encrypt(value);
 	}
 
-	public String decrypt(final String value, final String password) {
-		return stringEncryptor(password).decrypt(value);
+	public static String encrypt(final String value, final String password, final String algorithm) {
+		return stringEncryptor(password, algorithm).encrypt(value);
 	}
 
-	private StringEncryptor stringEncryptor(final String password) {
+	public static String decrypt(final String value, final String password, final String algorithm) {
+		return stringEncryptor(password, algorithm).decrypt(value);
+	}
+
+	private static StringEncryptor stringEncryptor(final String password, final String algorithm) {
 		SimpleStringPBEConfig config = new SimpleStringPBEConfig();
 		config.setPassword(password);
 		StandardPBEStringEncryptor encryptor = new StandardPBEStringEncryptor();
-		encryptor.setAlgorithm("PBEWithHmacSHA512AndAES_128");
+		encryptor.setAlgorithm(algorithm);
 		encryptor.setIvGenerator(new RandomIvGenerator());
 		encryptor.setConfig(config);
 		return encryptor;