+++ src/main/java/org/thermoweb/intellij/plugin/encrypt/UncryptStringAction.java
@@ -1,6 +1,9 @@
 package org.thermoweb.intellij.plugin.encrypt;
 
-import org.apache.commons.lang.StringUtils;
+import java.util.Map;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
 import org.jetbrains.annotations.NotNull;
 
 import com.intellij.openapi.actionSystem.AnAction;
@@ -11,26 +14,39 @@ import com.intellij.openapi.editor.Caret;
 import com.intellij.openapi.editor.Document;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.ui.Messages;
+
+import static org.thermoweb.intellij.plugin.encrypt.CipherInformationsDialog.ALGORITHM_FIELD_NAME;
+import static org.thermoweb.intellij.plugin.encrypt.CipherInformationsDialog.ENCAPSULATE_FIELD_NAME;
+import static org.thermoweb.intellij.plugin.encrypt.CipherInformationsDialog.PASSWORD_FIELD_NAME;
 
 public class UncryptStringAction extends AnAction {
-	private final CipherUtils cipherUtils = new CipherUtils();
+	private final Pattern pattern = Pattern.compile("ENC\\((.*)\\)");
 
 	@Override
 	public void actionPerformed(@NotNull final AnActionEvent event) {
-		String password = Messages.showPasswordDialog("Password : ", "Uncrypt String");
-		if (StringUtils.isEmpty(password)) {
+		CipherInformationsDialog dialog = new CipherInformationsDialog();
+		if (!dialog.showAndGet()) {
 			return;
 		}
+
+		Map<String, String> values = dialog.getValues();
 
 		Editor editor = event.getRequiredData(CommonDataKeys.EDITOR);
 		Project project = event.getRequiredData(CommonDataKeys.PROJECT);
 		Document document = editor.getDocument();
 		Caret primaryCaret = editor.getCaretModel().getPrimaryCaret();
+		if (primaryCaret.getSelectedText() == null) {
+			return;
+		}
+
+		Matcher matches = pattern.matcher(primaryCaret.getSelectedText());
+		final String selectedText = matches.find() && "true".equals(values.get(ENCAPSULATE_FIELD_NAME)) ?
+				matches.group(1) :
+				primaryCaret.getSelectedText();
 
 		WriteCommandAction.runWriteCommandAction(project,
 				() -> document.replaceString(primaryCaret.getSelectionStart(), primaryCaret.getSelectionEnd(),
-						cipherUtils.decrypt(primaryCaret.getSelectedText(), password)));
+						CipherUtils.decrypt(selectedText, values.get(PASSWORD_FIELD_NAME), values.get(ALGORITHM_FIELD_NAME))));
 
 		primaryCaret.removeSelection();
 	}